requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
llama-index
openai
python-dotenv
pypdf
chromadb
crewai
